+++ src/main/java/racingcar/model/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.car;
 
 import racingcar.strategy.MovingStrategy;
