+++ src/test/java/racingcar/model/CarTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.model.car.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
