+++ src/test/java/racingcar/model/RacingTest.java
@@ -2,6 +2,8 @@ package racingcar.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.car.Car;
+import racingcar.model.racing.Racing;
 
 import java.util.List;
